chore(frontend): split user context hook
