+++ frontend/src/lib/UserProvider.tsx
@@ -1,18 +1,6 @@
-import { createContext, useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import type { ReactNode } from 'react';
-
-export type User = {
-  id: string;
-  email?: string;
-  openaiKey?: string;
-  binanceKey?: string;
-  binanceSecret?: string;
-} | null;
-
-const UserContext = createContext<{ user: User; setUser: (u: User) => void }>({
-  user: null,
-  setUser: () => {},
-});
+import { UserContext, type User } from './user-context';
 
 export function UserProvider({ children }: { children: ReactNode }) {
   const [user, setUser] = useState<User>(() => {
@@ -26,8 +14,4 @@ export function UserProvider({ children }: { children: ReactNode }) {
   }, [user]);
 
   return <UserContext.Provider value={{ user, setUser }}>{children}</UserContext.Provider>;
-}
-
-export function useUser() {
-  return useContext(UserContext);
 }
