+++ frontend/src/lib/useUser.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { UserContext } from './user-context';
+
+export function useUser() {
+  return useContext(UserContext);
+}
