+++ frontend/src/lib/user-context.ts
@@ -1,0 +1,14 @@
+import { createContext } from 'react';
+
+export type User = {
+  id: string;
+  email?: string;
+  openaiKey?: string;
+  binanceKey?: string;
+  binanceSecret?: string;
+} | null;
+
+export const UserContext = createContext<{ user: User; setUser: (u: User) => void }>({
+  user: null,
+  setUser: () => {},
+});
